Set up C/C++, document process used so far
I'm downloading stdc++.h. It appears in the #include
directive from the C++ challenge fetched today. This is
an extension for the GNU GCC compiler that doesn't
exist in Apple's Clang compiler. Online, it's been
suggested to just download it and put into a
/include ¹ ². I'm downloading it inside the repository
here because I don't want to pollute the global
environment. I downloaded stdc++.h from the GCC
repository ³.

¹ https://apple.stackexchange.com/a/367285
² https://stackoverflow.com/questions/28994148
³ https://github.com/gcc-mirror/gcc/blob/master/libstdc%2B%2B-v3/include/precompiled/stdc%2B%2B.h